requirements.txt
# Kept in sync with pyproject.toml dependencies. Prefer: uv sync
feedparser>=6.0.0
requests>=2.28.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.27.0
google-genai>=0.6.0
tqdm
newspaper3k
